M2: pointer-truncation sweep (37 sites) + CI tripwire
Widen 37 pointer-as-int truncation sites that the frozen 32-bit layout
was hiding, and add a regression tripwire so new ones cannot land.

Sweep (behavior-identical on 32-bit, correct on 64-bit):
- 34 same-object pointer comparisons `(int)p < (int)&q` -> `(uintptr_t)`
  (loop bounds / conditionals across game/qcommon/server/client/cgame/
  gfx_d3d/sound/xanim). Signed->unsigned is result-identical for pointers
  into the same object (they never straddle 0x80000000) and matches the
  existing com_memory.cpp uintptr_t convention.
- 3 true address-alignment masks: db_stream.cpp:84 (live M5 stream-alloc
  path), cl_main.cpp:625 (skeleton mem), xmodel_load_obj.cpp:237 ->
  `& ~(uintptr_t)0xF` / `(uintptr_t)`.
Left untouched: the 24 semantic-rewrite sites (address stored into a
32-bit field/slot — that is M4 ABI work) and the 11 legitimate
size/index masks; both are recorded in the burndown allowlist.

Tripwire (tests/pointer_truncation_test.cmake + .allow):
- CMake script-mode CTest cloned from the headless-include-debt harness;
  zero compiler dependency, so it runs on all five portable-tests CI
  runners for free. Catches new (int)&/(uint32_t)&/(unsigned int)&/(DWORD)&
  casts and 0xFFFFF000/0xFFFFFFF0 address masks; FATALs on new debt or a
  stale allowlist entry (forcing burndown as ABI work removes sites).
- Seeded with the 37 remaining known-debt sites; portable suite is 5/5.

The P3 bare-hex-sizeof-assert funnel is deferred to M4 (needs the ABI
layout-header split). Engine TUs cannot be compiled on Linux, so the
MSVC compile of the 37 edits is confirmed by the Windows x86 CI leg.

+++ src/cgame/cg_compass.cpp
@@ -2217,7 +2217,7 @@ void CG_CompassDrawPlayerPointers_SP(
                 } while (v19);
             }
             ++objectives;
-        } while ((int)objectives < (int)cgArray[0].targets);
+        } while ((uintptr_t)objectives < (uintptr_t)cgArray[0].targets);
     }
 }
 #endif
