+++ src/client/cl_scrn.cpp
@@ -428,7 +428,7 @@ LABEL_20:
             R_SaveCubemapShot((char*)v29, v27, v10, v11);
             ++v28;
             ++v27;
-        } while ((int)v28 < (int)&szShotName[2]);
+        } while ((uintptr_t)v28 < (uintptr_t)&szShotName[2]);
     }
     else
     {
